Add bound_rect to hold X, Y, Width, Height info
This info was being held in multiple places (offsets, screen_info,
space_info, window_info, etc) and this commit consolidates all of
them into a new struct which has this info.

This also frees up a level of encapsulation, since tree/node functions
no longer need anything from the Screen structs. Previously they needed
the Screen to create the RootNodeContainer, because the Screen held the
dimensions. Now, the Space is created with a bound_rect with baked in
padding information. We can pass that bound_rect and the
RootNodeContainer will be sized to fit this attribute of the Space.

TODO -- Update the header file documentation for display et al.