Update GetElementByID with more templates.
Convert GetElementByID to GetElementByKey, and add templates to allow
the function to operate on templated data in the Node, use a templated
key for attribute lookup, and return a templated match if found.

This allows the function to search the Element for arbitrary matches.

The existing search functions which use a function of the Element can
accomplish a lot already, but this is mainly for finding things in the
Element's lower-level attributes.

Most notably in this project, it'll be a way to automatically search the
window_info cache during regular traversal.

This saves time when doing things like finding a Window by WID in the
Tree. Without this function, you'd need to collect all of the Elements,
and then process each Window in each Element to find the WID.
This allows the search to take place during regular tree lookup:

<return attribute> of <Window> in <WindowStack> of Node if
<Window->WID> matches <Key WID>

The attribute can be anything -- a WindowRef, the window_info* itself,
etc.

Similarly, the Key can be anything as well (Name, Owner, Role, etc).
This means that we need to define helper functions to do these lookups,
or we can return the Element itself and do the lookup in the calling
scope, which can be tedious if done repeatedly throughout the codebase.

It's probably worthwhile to convert this structure into a Collection
instead of just a Find, or determine a way to do both of these
together... after all a find is just collecting the first element, but
it's faster because you stop after a certain one.

(Maybe the Get and Collect can both call a function which does collection
but stops after a certain number of matches? Get would stop after 1
match, and Collect would stop after (inf) matches, or tree searched.)

Also updated RemoveElementFromTree to return a boolean success
indicator.